feat: 입력받은 자동차 이름이 5자를 초과하였을 때 예외를 출력한다.
입력받은 자동차 이름이 5자를 초과하였을 때 예외를 출력한다.
