+++ src/main/java/racingcargame/domain/Mileage.java
@@ -1,3 +1,5 @@
+package racingcargame.domain;
+
 import java.util.Objects;
 
 public class Mileage {
