+++ src/main/java/racingcargame/exception/NameLengthExceededException.java
@@ -1,0 +1,7 @@
+package racingcargame.exception;
+
+public class NameLengthExceededException extends IllegalArgumentException {
+    public NameLengthExceededException(String messageExceededNameLength) {
+        super(messageExceededNameLength);
+    }
+}
