+++ src/main/java/racingcargame/utility/Utility.java
@@ -1,3 +1,5 @@
+package racingcargame.utility;
+
 import java.util.Random;
 
 public class Utility {
