+++ src/test/java/racingcargame/utility/UtilityTest.java
@@ -1,3 +1,6 @@
+package racingcargame.utility;
+
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,6 +9,7 @@ class UtilityTest {
 
 
     @Test
+    @DisplayName("랜덤으로 발생한 정수가 0 이상 10 미만이다.")
     void generateSingleDigitIntegerTest() {
         // given, when
         Integer randomSingleDigitInteger = Utility.generateRandomSingleDigitInteger();
